feat(core): Move queries into its own module
