+++ geekorm-core/src/queries/mod.rs
@@ -1,0 +1,9 @@
+//! Queries Module
+
+/// The QueryBuilder Module
+pub mod builder;
+/// The Query Module
+pub mod query;
+
+pub use builder::QueryBuilder;
+pub use query::Query;
